# projects_manager_backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/students/add-request:
    post:
      summary: 'New add request'
      operationId: newAddRequest
      description: 'This endpoint allows students to send add requet to other students to join a group .'
      parameters: []
      responses: {  }
      tags:
        - 'Add Requests'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                group_id:
                  type: integer
                  description: 'The id of group .'
                  example: 7
                reciver_student_id:
                  type: integer
                  description: 'The id of the student who will recive the request .'
                  example: 7
              required:
                - group_id
                - reciver_student_id
  /api/students/update-add-request:
    post:
      summary: 'Update add request'
      operationId: updateAddRequest
      description: 'This endpoint allows students to accept or reject add requet from other students to join a group .'
      parameters: []
      responses: {  }
      tags:
        - 'Add Requests'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add_request_id:
                  type: integer
                  description: 'The id request .'
                  example: 19
                new_status:
                  type: integer
                  description: 'New status value .'
                  example: 1
                  enum:
                    - 0
                    - 1
              required:
                - add_request_id
                - new_status
...
  /api/admins/auth/register:
    post:
      summary: 'Admin Register'
      operationId: adminRegister
      description: 'Make new account for admins .'
      parameters: []
      responses: {  }
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: "Admin's first name ."
                  example: Ali
                last_name:
                  type: string
                  description: "Admin's last name ."
                  example: Mohammad
                email:
                  type: string
                  description: "Admin's email ."
                  example: alimo@gmail.com
                password:
                  type: string
                  description: "Admin's password ."
                  example: '4&svd&f'
                section:
                  type: string
                  description: "Admin's section ."
                  example: software
              required:
                - first_name
                - last_name
                - email
                - password
                - section
      security: []
  /api/admins/auth/login:
    post:
      summary: 'Admin login'
      operationId: adminLogin
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "Admin's email ."
                  example: alimo@gmail.com
                password:
                  type: string
                  description: "Admin's password ."
                  example: '4&svd&f'
                section:
                  type: string
                  description: "Admin's section ."
                  example: software
              required:
                - email
                - password
                - section
      security: []
...
  /api/doctors/auth/login:
    post:
      summary: 'Doctor login'
      operationId: doctorLogin
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "Doctor's email ."
                  example: alimo@gmail.com
                password:
                  type: string
                  description: "Doctor's password ."
                  example: '4&svd&f'
                section:
                  type: string
                  description: "Doctor's section ."
                  example: software
              required:
                - email
                - password
                - section
      security: []
...
  /api/students/auth/register:
    post:
      summary: 'Student Register'
      operationId: studentRegister
      description: 'Make new account for student .'
      parameters: []
      responses: {  }
      tags:
        - Student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ssn:
                  type: string
                  description: "Student's ssn ."
                  example: '10234567891'
                university_id:
                  type: string
                  description: "Student's university ID ."
                  example: '1452'
                first_name:
                  type: string
                  description: "Student's first name ."
                  example: Ali
                mid_name:
                  type: string
                  description: "Student's mid name ."
                  example: Hani
                last_name:
                  type: string
                  description: "Student's last name ."
                  example: Mohammad
                email:
                  type: string
                  description: "Student's email ."
                  example: alimo@gmail.com
                password:
                  type: string
                  description: "Student's password ."
                  example: '4&svd&f'
                section:
                  type: string
                  description: "Student's section ."
                  example: software
                year_of_study:
                  type: integer
                  description: "Student's year of study ."
                  example: 4
              required:
                - ssn
                - university_id
                - first_name
                - mid_name
                - last_name
                - email
                - password
                - section
                - year_of_study
      security: []
  /api/students/auth/login:
    post:
      summary: 'Student Login'
      operationId: studentLogin
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: "Student's email ."
                  example: alimo@gmail.com
                password:
                  type: string
                  description: "Student's password ."
                  example: '4&svd&f'
                section:
                  type: string
                  description: "Student's section ."
                  example: software
                year_of_study:
                  type: integer
                  description: "Student's year of study ."
                  example: 4
              required:
                - email
                - password
                - section
                - year_of_study
      security: []
...